+++ src/main/java/io/github/siloonk/protocol/data/GameState.java
@@ -1,4 +1,4 @@
-package io.github.siloonk.protocol;
+package io.github.siloonk.protocol.data;
 
 public enum GameState {
     HANDSHAKING,
