+++ src/main/java/io/github/siloonk/protocol/packets/status/PingResponsePacket.java
@@ -1,7 +1,7 @@
 package io.github.siloonk.protocol.packets.status;
 
 import io.github.siloonk.protocol.Packet;
-import io.github.siloonk.protocol.PacketDirection;
+import io.github.siloonk.protocol.data.PacketDirection;
 import io.github.siloonk.protocol.streams.PacketInputStream;
 import io.github.siloonk.protocol.streams.PacketOutputStream;
